use director logging
