+++ components/system-broker/internal/osb/bind_create.go
@@ -22,8 +22,8 @@ import (
 
 	schema "github.com/kyma-incubator/compass/components/director/pkg/graphql"
 
+	"github.com/kyma-incubator/compass/components/director/pkg/log"
 	"github.com/kyma-incubator/compass/components/system-broker/internal/director"
-	"github.com/kyma-incubator/compass/components/system-broker/pkg/log"
 	"github.com/pivotal-cf/brokerapi/v7/domain"
 	"github.com/pivotal-cf/brokerapi/v7/domain/apiresponses"
 	"github.com/pkg/errors"
