+++ components/system-broker/internal/osb/bind_get.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/kyma-incubator/compass/components/system-broker/internal/director"
 
-	"github.com/kyma-incubator/compass/components/system-broker/pkg/log"
+	"github.com/kyma-incubator/compass/components/director/pkg/log"
 	"github.com/pivotal-cf/brokerapi/v7/domain"
 	"github.com/pivotal-cf/brokerapi/v7/domain/apiresponses"
 	"github.com/pkg/errors"
