+++ components/system-broker/internal/osb/instance_create.go
@@ -18,7 +18,8 @@ package osb
 
 import (
 	"context"
-	"github.com/kyma-incubator/compass/components/system-broker/pkg/log"
+
+	"github.com/kyma-incubator/compass/components/director/pkg/log"
 	"github.com/pivotal-cf/brokerapi/v7/domain"
 )
