+++ components/system-broker/internal/osb/instance_update.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/kyma-incubator/compass/components/system-broker/pkg/log"
+	"github.com/kyma-incubator/compass/components/director/pkg/log"
 	"github.com/pivotal-cf/brokerapi/v7/domain"
 )
