+++ components/system-broker/pkg/graphql/client.go
@@ -19,7 +19,7 @@ package graphql
 import (
 	"context"
 
-	"github.com/kyma-incubator/compass/components/system-broker/pkg/log"
+	"github.com/kyma-incubator/compass/components/director/pkg/log"
 	"github.com/machinebox/graphql"
 	"github.com/pkg/errors"
 )
