+++ components/system-broker/pkg/server/server.go
@@ -26,7 +26,7 @@ import (
 	"time"
 
 	"github.com/gorilla/mux"
-	"github.com/kyma-incubator/compass/components/system-broker/pkg/log"
+	"github.com/kyma-incubator/compass/components/director/pkg/log"
 	"github.com/kyma-incubator/compass/components/system-broker/pkg/panic_recovery"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
@@ -39,7 +39,7 @@ type Server struct {
 	shutdownTimeout time.Duration
 }
 
-func New(c *Config, service log.UUIDService, middlewares []mux.MiddlewareFunc, routesProvider ...func(router *mux.Router)) *Server {
+func New(c *Config, middlewares []mux.MiddlewareFunc, routesProvider ...func(router *mux.Router)) *Server {
 	s := &Server{
 		shutdownTimeout: c.ShutdownTimeout,
 		routesProvider:  routesProvider,
@@ -58,7 +58,7 @@ func New(c *Config, service log.UUIDService, middlewares []mux.MiddlewareFunc, r
 	router.HandleFunc(c.RootAPI+"/debug/pprof/symbol", pprof.Symbol)
 	router.HandleFunc(c.RootAPI+"/debug/pprof/trace", pprof.Trace)
 
-	router.Use(log.RequestLogger(service))
+	router.Use(log.RequestLogger())
 	router.Use(panic_recovery.NewRecoveryMiddleware())
 
 	for _, m := range middlewares {
