+++ components/system-broker/pkg/signal/signal.go
@@ -22,7 +22,7 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/kyma-incubator/compass/components/system-broker/pkg/log"
+	"github.com/kyma-incubator/compass/components/director/pkg/log"
 )
 
 // HandleInterrupts handles process signal interrupts
